pillow_approach/requirements.txt
pillow==10.2.0
opencv-python==4.5.3
ttkbootstrap==1.10.1